Transitioning Contact page to use HAML templates
Using the openclerk/pages component

Issue #399
