+++ templates/contact.haml
@@ -1,0 +1,4 @@
+%h1
+  = t("Contact us")
+
+!= requireTemplate("inline_contact");
