--- templates/contact.php
@@ -1,5 +1,0 @@
-<h1><?php echo ht("Contact Us"); ?></h1>
-
-<?php
-require_template('inline_contact');
-?>
